Add remote registry page routed tabs
