+++ cypress/e2e/hub/remote-registries.cy.ts
@@ -101,7 +101,7 @@ describe('Remote Registry', () => {
     cy.get('[data-cy="name"]').type(remoteRegistryName);
     cy.get('[data-cy="url"]').type(RemoteRegistry.remoteURL);
     cy.get('[data-cy="Submit"]').click();
-    cy.url().should('include', `remote-registries/details/${remoteRegistryName}`);
+    cy.url().should('include', `remote-registries/${remoteRegistryName}/details`);
     cy.contains('Remote registries').click();
     cy.url().should('include', RemoteRegistry.url);
     cy.filterTableBySingleText(remoteRegistryName);
@@ -134,14 +134,7 @@ describe('Remote Registry', () => {
       cy.url().should('include', `remote-registries/${remoteRegistryName}/edit`);
       cy.get('[data-cy="url"]').clear().type(RemoteRegistry.remoteURL);
       cy.clickButton(/^Save remote registry$/);
-      cy.clickButton(/^Clear all filters$/);
-      cy.filterTableBySingleText(remoteRegistryName);
-      cy.contains('tr', remoteRegistryName).within(() => {
-        cy.contains('td', remoteRegistryName).within(() => {
-          cy.get('a').click();
-        });
-      });
-      cy.url().should('include', `remote-registries/details/${remoteRegistryName}`);
+      cy.url().should('include', `remote-registries/${remoteRegistryName}/details`);
       cy.get('[data-cy="name"]').should('contain', remoteRegistryName);
       cy.get('[data-cy="url"]').should('contain', RemoteRegistry.remoteURL);
